// frontend/src/api/model/githubCom4H1RZooraInternalDomainOfflineRoom.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainClass } from './githubCom4H1RZooraInternalDomainClass';
import type { GithubCom4H1RZooraInternalDomainClassSession } from './githubCom4H1RZooraInternalDomainClassSession';
import type { GithubCom4H1RZooraInternalDomainMedia } from './githubCom4H1RZooraInternalDomainMedia';
import type { GithubCom4H1RZooraInternalDomainUser } from './githubCom4H1RZooraInternalDomainUser';

export interface GithubCom4H1RZooraInternalDomainOfflineRoom {
  attachments?: GithubCom4H1RZooraInternalDomainMedia[];
  class?: GithubCom4H1RZooraInternalDomainClass;
  class_id?: string;
  class_session?: GithubCom4H1RZooraInternalDomainClassSession;
  class_session_id?: string;
  created_at?: string;
  creator?: GithubCom4H1RZooraInternalDomainUser;
  creator_id?: string;
  description?: string;
  id?: string;
  organization_id?: string;
  published_at?: string;
  title?: string;
  updated_at?: string;
  view_count?: number;
}